requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.0
defusedxml==0.7.0rc1
Django==3.0.8
django-braces==1.14.0
django-cors-headers==3.4.0
django-extensions==3.0.5
django-oauth-toolkit==1.3.2
django-organizations==1.1.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
idna==2.10
kombu==4.6.11
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.10
vine==1.3.0
